requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
Django==5.0.7
django-apscheduler==0.6.2
django-cors-headers==4.4.0
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
joblib==1.4.2
# mysqlclient==2.2.4
numpy==2.0.0
packaging==24.1
pillow==10.4.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzlocal==5.2
whitenoise==6.7.0
